+++ src/sensor_parser/package.xml
@@ -2,12 +2,12 @@
 <package format="2">
   <name>sensor_parser</name>
   <version>0.0.0</version>
-  <description>The sensor_parser package</description>
+  <description>Parse the sensor data and publish to topics</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="oshadha@todo.todo">oshadha</maintainer>
+  <maintainer email="sflatter@andrew.cmu.edu">Sam Flattery</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
