+++ src/sensor_parser/scripts/tags:
@@ -1,0 +1,12 @@
+!_TAG_FILE_SORTED	2	/0=unsorted, 1=sorted, 2=foldcase/
+Imu	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^from sensor_msgs.msg import Imu, NavSatFix$/;"	i	language:Python
+NavSatFix	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^from sensor_msgs.msg import Imu, NavSatFix$/;"	i	language:Python
+parse_channel	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^def parse_channel(data_string, channel_number):$/;"	f	language:Python
+parse_GPS	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^def parse_GPS(data_string):$/;"	f	language:Python
+parse_IMU	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^def parse_IMU(data_string):$/;"	f	language:Python
+parse_sensors	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^def parse_sensors(ser):$/;"	f	language:Python
+Quaternion	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^from geometry_msgs.msg import Quaternion$/;"	i	language:Python
+rospy	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^import rospy$/;"	i	language:Python
+ser	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^    ser = serial.Serial()$/;"	v	language:Python
+serial	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^import serial$/;"	i	language:Python
+UInt32	/Users/samflattery/dev/DuneBuggyROS/src/sensor_parser/scripts/sensor_parser.py	/^from std_msgs.msg import UInt32$/;"	i	language:Python
